requirements.txt
# MkDocs and extensions
mkdocs>=1.5.0
mkdocs-material>=9.4.0
mkdocs-minify-plugin>=0.7.0
pymdown-extensions>=10.3.0
mkdocs-mermaid2-plugin>=1.2.0

# Additional useful plugins
mkdocs-git-revision-date-localized-plugin>=1.2.0
mkdocs-git-committers-plugin-2>=1.2.0
mkdocs-awesome-pages-plugin>=2.9.0
mkdocs-include-markdown-plugin>=6.0.0
mkdocs-macros-plugin>=1.0.0

# Link validation and quality checks
mkdocs-roamlinks-plugin>=0.3.0
mkdocs-link-marker>=0.2.0
pyyaml>=6.0